requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
ansicolors==1.1.8
ansiwrap==0.8.4
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.2.0
colorama==0.4.6
console-menu==0.8.0
frozenlist==1.4.0
idna==3.4
iniconfig==2.0.0
meraki==1.34.0
multidict==6.0.4
packaging==23.1
pluggy==1.2.0
pytest==7.4.0
requests==2.31.0
six==1.16.0
textwrap3==0.9.2
tomlkit==0.11.8
urllib3==2.0.3
yarl==1.9.2
